backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
python-dotenv
psycopg2-binary
Pillow
